+++ tests/test.php
@@ -33,5 +33,6 @@ include_once("utilityPayment.php");
 include_once("utilityPaymentLog.php");
 include_once("webhook.php");
 include_once("event.php");
+include_once("paymentRequest.php");
 
 echo "\n\nAll tests concluded\n\n";